More graphing stuff
The Graph wrapper now generates a checkbox for each line added which
controls the visibility of that line. The text color of the checkbox is
set to the same color as the corresponding line, and the text is set to
the name of the loss that the line graphs.

Also fixed the bug where graph line visibility updating was limited to
the update rate of the trainer worker.