More Graphing and UI updates
Changed the graphing class again. I thought it would be cool to graph
the iteration time on the performance graph but ironically, I didn't
consider the performance overhead updating the graphs that frequently.
After playing with it, I don't think the data it provides is useful
enough for the cost as it is now or the time to do it correctly. The end
result of this though is that all lines now keep track of their own
steps which I actually think was a smart change regardless.

Also made some more UI updates. Started adding the G and D architecture
options to the UI. Finally figured out the stylsheet inheritance issue
as well so I updated some other widget styles now that it's possible to
cascade them on the settings bar.