Bug fix and ONNXConverter changes
Fixed a bug in the preview button linking function in utility_panel. I hadn't noticed that
a previous refactor had introduced a late binding issue in UtilityPanel._init_utilities_buttons.
Fixed by adding a wrapper function around the lambda for the dry_run calls.

The ONNXConverter class also now takes a config (of any standard type apart from NoneType)
for init. This change was made because, while I was writing the docs for the ONNXConverter, I
realised 2 things:

1.) The ONNXConverter (in it's previous state) was not fun to work with outside of the Toolbox.
2.) It also had not been updated to accept the new G architecture values from the config.

It is still not perfect, but it's a bit better now than it was. In response to this, though,
UtilityPanel.convert_to_onnx now also does a quick check after it validates the experiment dir
to find the latest train config in the directory automatically to pass to the ONNXConverter.